+++ au3/modules/import-export/RegisterExportPlugins.h
@@ -1,0 +1,7 @@
+/*
+* Audacity: A Digital Audio Editor
+*/
+
+#pragma once
+
+void RegisterExportPlugins();
