+++ au3/modules/import-export/mod-ffmpeg/FFmpeg.cpp
@@ -55,10 +55,6 @@ void FFmpegStartup()
         //          \nbut this time Audacity failed to load it at startup. \
         //          \n\nYou may want to go back to Preferences > Libraries and re-configure it."),
         //!             XO("FFmpeg startup failed")
-    } else {
-        // calling it manually because static initialization in
-        // FFmpegImportPlugin.cpp doesn't work for some reason
-        FFmpegImportPlugin();
     }
 }
