+++ buildscripts/cmake/SetupDependencies.cmake
@@ -65,6 +65,7 @@ endfunction()
 populate(wxwidgets "wxwidgets/3.1.3.6")
 populate(expat "expat/2.0.5")
 populate(portaudio "portaudio/19.7.0")
+populate(libmp3lame "libmp3lame/3.100")
 
 if (NOT OS_IS_LIN)
     populate(zlib "zlib/1.2.13")
